Separated the control layer between Mover and PlayerController scripts
