+++ Assets/Game/Characters/Player.prefab
@@ -10,6 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 3828610826871390740}
   - component: {fileID: 3828610826871390741}
+  - component: {fileID: 1904542395686059160}
   - component: {fileID: 3828610826871390762}
   - component: {fileID: 3828610826871390763}
   m_Layer: 0
@@ -57,6 +58,18 @@ NavMeshAgent:
   m_BaseOffset: 0
   m_WalkableMask: 4294967295
   m_ObstacleAvoidanceType: 4
+--- !u!114 &1904542395686059160
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3828610826871390736}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2733841f7df89f9458a0d4fe65d74c8a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &3828610826871390762
 MonoBehaviour:
   m_ObjectHideFlags: 0
