+++ Assets/Scenes/Sandbox.unity
@@ -2085,3 +2085,60 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 214761b5faf78a543b6c473572316851, type: 3}
+--- !u!1001 &7339191301420535424
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_RootOrder
+      value: 4
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.416
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 4.58
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5111502693673952557, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8569139533916603429, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
+      propertyPath: m_Name
+      value: Enemy
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 8bc0b564e269e624b9cd149148b43eb9, type: 3}
