+++ Assets/Scripts/Control.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f4f5aba9375db8846a04dc2ebe4a0a95
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
