+++ Assets/Scripts/Control/PlayerController.cs
@@ -1,0 +1,36 @@
+using UnityEngine;
+using RPG.Movement;
+
+namespace RPG.Control
+{
+    [RequireComponent(typeof(Mover))]
+    public class PlayerController : MonoBehaviour
+    {
+        private Mover myMover;
+
+        private void Awake()
+        {
+            myMover = GetComponent<Mover>();
+        }
+
+        private void Update()
+        {
+            if (Input.GetMouseButton(0))
+            {
+                MoveToCursor();
+            }
+        }
+
+        private void MoveToCursor()
+        {
+            Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
+            RaycastHit hit;
+            bool hasHit = Physics.Raycast(ray, out hit);
+
+            if (hasHit)
+            {
+                myMover.MoveTo(hit.point);
+            }
+        }
+    }
+}
